Fix: sidebar refactor name files
